[#3] Make the code compatible with core v0.2.0
Before doing any more work, we need to update the code so it is compatible v0.2.0.

First of all, update the `src2` sample so the syntax is compatible.
Then, we move the `ScalaASTBuilder` class to a package `builder`.
Because of that, we can split the builder in more specific traits.

We do not really change the logic of the built code.
Most of it is moved code.
But we need to adapt with the new model.

The building change has an immediate impact into the generator plugin.
First of all, the notion of file disappeared of the AST.
So the generator must chose its own strategy to create its files.
The one chosen for the first step was to create only package object.

We stopped the changes when we had a compilable code.

This commit do the following:

* Correct sample with the new syntax
* Update `Verification` so it accepts a overriding message
* Move builder into its own package and split it
* Make the code compatible v0.2.0
* Adapt the generator plugin