Improve the validation system and add `List` and `Option` cases
If we want to validate our model, we encounter a strange issue.
`Option`al values and `List`s are not deeply checked.

To enable this feature, a refactoring of `Verification` model was needed.
We simplify the `Verification` concrete classes and add more classes.
Each concrete class do one and only one thing.
We use a mix of classes to compose all verifications.

This commit does the following:

* Improve the validation system
* In case of `Option` and `List`, create a deep `Verification`