Enable JSON support
Because in lot of cases, the model needs to be serialized into JSON,
we provide implicit conversions into json.

In scala, there is several libraries to use JSON.
We start here with spray-json (used in akka http).

If wanted, it is possible to directly validate the read value (configurable).

This commit does the following:

* Create the `JsonSpraySupport` in case json/spray support is enabled
* Create the `JsonBuilder` trait to build json parsers with a strategy pattern to choose the format
* Include the json statements for defined types
* Improve scala AST to have `implicit val` and add some constructors to simplify code
* Improve the `Resource` class to navigate through files*
* Correct an issue with `ListVerification` and `OptionVerification` when there is no nested `Verification`
* Improve `EndToEndSpec` and other test helpers to test codes with specific *scalamodel* configuration
* Create tests for json spray support