Correct `JsonSpraySupport` imports and `BigDecimal` implicit conversion
For the first error, it is just a copy/paste error when writing the code.
Because manual testing generated the types, import errors were not found out.

For the second, it was encountered when creating the sum function.
The scala implicit system could not determined `0` as `BigDecimal`,
giving a scala compilation error.

This commit does the following:

* Remove unwanted imports in `JsonSpraySupport`
* Force the use of `BigDecimal` when using a `NumberValue`