Do not duplicate validation for alias types
Testing real cases of the compiler,
we found that attributes with an alias types have alias type verifications duplicated.

We do not want to check twice the same thing so we remove the second call.

Also, during development of this correction,
we improve the test system so it will be more easier to understand errors.

This commit does the following:

* Do not include inheritance verifications for alias type because already done with alias type specific verification
* Create a test case to check the development
* Create a matcher to check if generated root is valid and pretty print it to see more efficiently differences