Add tests of native classes
Before going further with verification as a tree,
we add tests in native classes.

Firstly, we need our build system to consider our resources as test sources.
It provides us a way to test them and keep them as resources for production use.

This commit also add ScalaCheck to use property-based testing.

We limit our tests to `Validation` and `Verification`
because they are both in the scope of the next evolution.

Doing these tests, we encounter a bug with `OptionVerification` so we correct it.
Also, we remove unused functions and simplify code where possible.

This commit does the following:

* Add scalacheck to build system
* Add native resources as test sources
* Verify the behavior of `Validation`
* Verify the nominal behavior of `Verification`
* Add a more complete sample (behavior) of `Verification`: *Season*
* Simplify code for `Validation` and `Verification`