Verifications now give the path to the invalid field
Because we need to know which field is invalid,
we update the `Verification` and `Validation` system to specify it.

To specify the field, we use a string path as: `f1.f2[0].f3` (as javascript syntax).

The change will be done in several steps.
This one focuses on the native API.

It is a breaking change of the current version of the code.

This commit does the following:

* Update `Invalid` to receive an `Error(path, messages)` instead of messages only
* Update `Verification` to consider the updated `Invalid`
* Correct tests to show how it works
* Add `Season.options` in `VerificationSeasonSamplesSpec` for deep validation
* Add another test into `VerificationSeasonSamplesSpec` to test this deep validation
* Restructure `VerificationSeasonSamplesSpec` to clean the code