Improve the json compatibility
With the current improvement to have fields alongside error messages,
we can improve the compatibility with json formats.

For `JsonSpraySupport`, there is no mean to define an error for a specific path.
We create a custom `Exception` so the developer will be able to match it
and to transform it as s·he want.

This commit does the following:

* `JsonPlaySupport`: Link error messages to their respective JsPath
* `JsonSpraySupport`: Create `ValidationDeserializationException` to throw a custom exception for validation issues