Correct the deep validation of json parsers
When testing json parsers in real use cases, we encountered a strange behavior.
When there was an error from validations in nested type,
only the error into the nested type were returned.

This commit resolves this issue by forcing the building of the type
without validations and validate the whole aggregate at last.

In case of Json with validation, json builders now have two parsers:
an implicit one and a raw one.
The raw one is used when the type is used in another one.
We import the nested type raw parser as implicit to build the main raw parser.
It is only for external uses that the implicit parser (with validation) will be used.

This commit does the following:

* Change the json configuration, to have validations as a boolean instead of an enum
* Json builders now use `ScalaModelBuilder` so we change their design
* Json builders have two parsers: a raw one and an implicit one (see above)
* Update `ScalaAST.Val` so we can have `implicit` local `val`
* Update the body generation so when there is several statements, we use braces
* Remove unused `TypeBuilder.generateTag` (old strategy removed during 0.2.0 development)
* Change sample `src2/person.def` to have a nested type