Improve and correct tests after passing to v0.3.0
The upgrade to v0.3.0 broke all tests.
We need to correct them so they can still be used.

We also take this opportunity to improve tests.
Until now, we read a definiti file and test a scala AST.
The understanding of tests was quite complex and long to write.
Instead, we now test the generator from definiti file to scala file.

This commit does the following:

* Improve the test system
* Restructure test files
* Correct `AliasTypeBuilder.generateVerification` because tests explicit a bug
* Correct tests according to new system
* Correct tests according to new AST